+++ render_block/base.py
@@ -1,6 +1,6 @@
 from typing import List, Optional, Tuple, Union
 
-from django.http import HttpRequest
+from django.http import HttpRequest, HttpResponse
 from django.template import Context, loader
 from django.template.backends.django import Template as DjangoTemplate
 
@@ -27,10 +27,13 @@ def render_block_to_string(
     Loads the given template_name and renders the given block with the given
     dictionary as context. Returns a string.
 
-        template_name
+    :param template_name:
             The name of the template to load and render. If it's a list of
             template names, Django uses select_template() instead of
             get_template() to find the template.
+    :param block_name: The name of the block to load.
+    :param context: The context dictionary used while rendering the template.
+    :param request: The request that triggers the rendering of the block.
     """
 
     # Like render_to_string, template_name can be a string or a list/tuple.
@@ -55,3 +58,19 @@ def render_block_to_string(
         raise UnsupportedEngine(
             "Can only render blocks from the Django template backend."
         )
+
+
+def render_block(
+    request: HttpRequest,
+    template_name: str,
+    block_name: str,
+    context: Optional[Context] = None,
+    content_type: Optional[str] = None,
+    status: Optional[int] = None,
+) -> HttpResponse:
+    """
+    Return an HttpResponse whose content is filled with the result of calling
+    render_block.render_block_to_string() with the passed arguments.
+    """
+    content = render_block_to_string(template_name, block_name, context, request)
+    return HttpResponse(content, content_type, status)